Move core part of Documentation layout into theme
