+++ plugins/gatsby-theme-iterative-docs/src/components/Documentation/Layout/SearchForm/index.tsx
